+++ compiler/src/main/kotlin/com/r0adkll/kimchi/processors/ContributesMultibindingSymbolProcessor.kt
@@ -7,7 +7,6 @@ import com.google.devtools.ksp.processing.SymbolProcessor
 import com.google.devtools.ksp.processing.SymbolProcessorEnvironment
 import com.google.devtools.ksp.processing.SymbolProcessorProvider
 import com.google.devtools.ksp.symbol.KSClassDeclaration
-import com.r0adkll.kimchi.HINT_MULTIBINDING_PACKAGE
 import com.r0adkll.kimchi.annotations.ContributesMultibinding
 import com.r0adkll.kimchi.annotations.ContributesMultibindingAnnotation
 import com.r0adkll.kimchi.util.KimchiException
@@ -17,7 +16,7 @@ import kotlin.reflect.KClass
 
 internal class ContributesMultibindingSymbolProcessor(
   env: SymbolProcessorEnvironment,
-) : HintSymbolProcessor(env, HINT_MULTIBINDING_PACKAGE) {
+) : HintSymbolProcessor(env) {
 
   @AutoService(SymbolProcessorProvider::class)
   class Provider : SymbolProcessorProvider {