+++ sample/androidApp/src/main/java/com/r0adkll/kimchi/restaurant/android/di/AndroidAppComponent.kt
@@ -27,7 +27,7 @@ abstract class AndroidAppComponent(
       packageName = application.packageName,
       debugBuild = BuildConfig.DEBUG,
       flavor = Flavor.Standard,
-      versionName = packageInfo.versionName,
+      versionName = packageInfo.versionName ?: "",
       versionCode = packageInfo.versionCode,
     )
   }